requirements.txt
python-pptx>=0.6.21
pypdf>=4.0.0
requests>=2.31.0
genanki>=0.13.1
rich>=13.7.0
google-generativeai>=0.8.0
flask>=3.0.0
